// src/modules/featureConfiguration/enum/featureConfigurationFeature.enum.ts
export enum FeatureConfigurationFeature {
  DisableReasons = 'disableReasons',
  Localization = 'Localization',
  Locations = 'locations',
  Contracts = 'contracts',
  Invoices = 'invoices',
  MyColleagues = 'my-colleagues',
  MyColleaguesForm = 'my-colleagues-form',
  StaffingRequests = 'staffing-requests',
  Invitation = 'invitation',
  Communications = 'communications',
  CreateJobOrderFormTemporary = 'createJobOrderFormTemporary',
  CreateJobOrderFormPermanent = 'createJobOrderFormPermanent',
  CreateJobOrderType = 'createJobOrderType',
  MainMenu = 'mainMenu',
  SupportedFileTypesForUpload = 'supportedFileTypesForUpload',
  OrderDetailsTimeline = 'orderDetailsTimeline',
  OrderDetailsListingStatuses = 'orderDetailsListingStatuses',
  OrderDetailsActions = 'jobOrderActions',
  CandidatesList = 'candidates-list',
  MailProvider = 'mailProvider',
  Cases = 'cases',
  JobRoleTemplate = 'jobRoleTemplate',
}