// src/seed/tenantSpecific/featureConfiguration/features/CreateJobOrderForm/enum/createJobOrderFormCellType.enum.ts
export enum CreateJobOrderFormCellType {
  Date = 'date',
  Input = 'input',
  Select = 'select',
  DatePicker = 'datePicker',
  Incrementer = 'incrementer',
  DayPicker = 'dayPicker',
  Languages = 'languages',
  AddMoreButton = 'addMoreButton',
  Certifications = 'certifications',
  TextArea = 'textArea',
  InterviewRequired = 'interviewRequired',
  UploadFiles = 'uploadFiles',
  EmptyCell = 'emptyCell',
  FileList = 'fileList',
  OrderName = 'orderName',
  OrderSalary = 'orderSalary',
  OrderWhichDays = 'orderWhichDays',
  OrderMap = 'orderMap',
  OrderInterviewRequired = 'orderInterviewRequired',
  OrderAdditionalInformation = 'orderAdditionalInformation',
  OrderFileList = 'orderFileList',
  OrderRole = 'orderRole',
  InputPlusMinus = 'inputPlusMinus',
  SectionLabel = 'sectionLabel',
  DateRange = 'dateRange',
  SingleDatePicker = 'singleDatePicker',
  CustomLabel = 'customLabel',
  InfoCard = 'infoCard',
  TextInformation = 'textInformation',
  TimePicker = 'timePicker',
  TextGuidanceInformation = 'textGuidanceInformation',
  Level = 'level',
  SimpleSelect = 'simpleSelect',
  ApproverContact = 'approverContact',
  ReportingToContact = 'reportingToContact',
  BillToContact = 'billToContact',
  PesRegistrationMessage = 'pesRegistrationMessage',
}